requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers>=2.7.0
duckduckgo-search==4.1.1
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
diskcache==5.6.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
lxml==4.9.3
html2text==2020.1.16
